Fix: 좋아요, 좋아요 취소 기능

+++ src/main/java/com/banbanmoomani/memilog/DAO/PostMapper.java
@@ -29,9 +29,9 @@ public interface PostMapper {
 
     void deletePostById(int postId);
 
-    void increaseLikeCount(@Param("post_id") int post_id);
+    void increaseLikeCount(@Param("post_id") Long post_id);
 
-    void decreaseLikeCount(@Param("post_id") int post_id);
+    void decreaseLikeCount(@Param("post_id") Long post_id);
 
     List<todayPostDTO> findTodayPost();
